adds golden path integration tests for both tcp and http
